scraper/requirements.txt
feedparser==6.0.11
openai==1.30.0
supabase==2.9.0
python-dotenv==1.0.1
trafilatura==1.12.2
requests==2.32.3
google-generativeai==0.8.3